Fix import in SPA app
